Create Cloud computing AWS IAM, EC2, S3
Overview 
This week we will get some hands-on experience with EC2 and S3. You have been using EC2 for a couple weeks now, but this assignment will dive deeper into its capabilities. 
One of the challenges of an assignment like this is that the AWS web console changes all the time without warning. Therefore, the instructions can’t be too tightly coupled to the AWS web interface. In other words, instead of telling you exactly which link to select or button to push on the AWS web console, you will need to make reasonable decisions as you work through the assignment.